chore: Move all server app setup to the app.py module
